+++ eds-aic-20230930/alibabacloud_eds_aic20230930/models/_reset_android_instances_in_group_request.py
@@ -10,11 +10,13 @@ class ResetAndroidInstancesInGroupRequest(DaraModel):
     def __init__(
         self,
         android_instance_ids: List[str] = None,
+        ignore_param_validation: bool = None,
         sale_mode: str = None,
         setting_reset_type: int = None,
     ):
         # The IDs of the cloud phone instances.
         self.android_instance_ids = android_instance_ids
+        self.ignore_param_validation = ignore_param_validation
         self.sale_mode = sale_mode
         self.setting_reset_type = setting_reset_type
 
@@ -29,6 +31,9 @@ class ResetAndroidInstancesInGroupRequest(DaraModel):
         if self.android_instance_ids is not None:
             result['AndroidInstanceIds'] = self.android_instance_ids
 
+        if self.ignore_param_validation is not None:
+            result['IgnoreParamValidation'] = self.ignore_param_validation
+
         if self.sale_mode is not None:
             result['SaleMode'] = self.sale_mode
 
@@ -41,6 +46,9 @@ class ResetAndroidInstancesInGroupRequest(DaraModel):
         m = m or dict()
         if m.get('AndroidInstanceIds') is not None:
             self.android_instance_ids = m.get('AndroidInstanceIds')
+
+        if m.get('IgnoreParamValidation') is not None:
+            self.ignore_param_validation = m.get('IgnoreParamValidation')
 
         if m.get('SaleMode') is not None:
             self.sale_mode = m.get('SaleMode')